Allow updating view backwards on catchup commit
When a node starts up and needs to catch-up from a long way back, it may
trigger view changes such that it is on a later view than the one the
blocks it commits were committed at. In this case, the node should be
able to correct itself and adopt the views in which the blocks were
committed.

Signed-off-by: Logan Seeley <seeley@bitwise.io>