Fix signer verification bug
This bug fixes a serious security issue caused by a bug in the way the
signers of messages and votes are verified.

Previously, when messages were received, the PbftMessage's signer_id was
compared with the ID of the node that sent the message; however, the
node that sent the message may not be the node that signed it, so this
is the wrong thing to compare it to; instead, it should be comparing the
PbftMessage's signer_id with the PeerMessageHeader's signer_id (this is
the ID that is used by the validator to verify the proper signing of the
consensus message).

Furthermore, when votes are verified for the consensus seal and view
changes, the IDs used to check membership and uniqueness are the IDs in
the PbftMessages themselves, and they were previously not compared with
the IDs that are used to actually sign the messages (again, indicated by
the signer_id of the ConsensusPeerMessageHeader). They should be
compared with the actual signing ID to be sure they match.

These changes break existing unit tests, since signing of messages in
these tests was just done with a random public public key that did not
match IDs in the or the IDs of the actual messages. Since these unit
tests are being replaced, they are not fixed but instead removed.

Signed-off-by: Logan Seeley <seeley@bitwise.io>