Fix ParsedMessage type vulnerability
This commit fixes a vulnerability that would allow a malicious node to
arbitrarily crash another node's PBFT engine.

When parsing messages from peers, nodes use the type indicated in the
header of the PeerMessage; however, this type field can be set
arbitrarily and could be set to a different value than the contained
PBFT message's type. Some of the accessors for ParsedMessages panic if
called on the wrong type of ParsedMessage (e.g. calling get_seal on a
ParsedMessage containing a NewView), so there must be a mechanism to
guard against a bad message triggering a panic.

This is fixed by simply checking and ensuring that the ParsedMessage's
type matches the type of the PeerMessage. An alternative would be to
remove the panics in the accessors, but those are useful for ensuring
that no programmer error occurs; we should be able to guarantee proper
parsing of messages, and once those messages are parsed they should only
be handled as the appropriate type.

Signed-off-by: Logan Seeley <seeley@bitwise.io>