Fix from_self value of ParsedMessages from peers
Previously, when a PeerMessage was parsed, it was assumed that the
message came from another node. However, it is possible for a node to
send a message back to the node that created the message, in which case
the ParsedMessage.from_self field should actually be true.

This commit fixes that assumption by explicitly checking if the
message's signer ID matches the node's own ID.

Signed-off-by: Logan Seeley <seeley@bitwise.io>