Fix terminology for PBFT member nodes (was peers)
Sawtooth already uses the term "peer" for static/dynamic peering, so the
PBFT doc was overloading this term. This PR changes "peer" to "(PBFT) member node"
and adds a glossary definition for "member node".

Note: The Sawtooth documentation will be fixed in a related-but-separate PR for
the hyperledger/sawtooth repo.

Signed-off-by: Anne Chenette <chenette@bitwise.io>