Bootstrap with Commit for chain head
When the whole network is starting "fresh" from a non-genesis block,
none of the nodes will have the `Commit` messages necessary to build the
consensus seal for the last committed block (the chain head). To
bootstrap the network in this scenario, all nodes will send a `Commit`
message for their chain head in two cases:

1. When the node starts, if it is already connected to other peers, it
will broadcast the commits to those peers
2. Whenever one of the PBFT members connects

When > 2f + 1 nodes have connected and received these `Commit` messages,
the nodes will be able to build a seal using the messages.

This commit includes a shell script to verify the ability to restart a
network with an existing chain using the adhoc docker compose files.
Also adds this test to CI.

Signed-off-by: Logan Seeley <seeley@bitwise.io>