Fix usage of the Ticker in the main engine loop
Fixes the usage of the `Ticker` in the main engine loop; this `Ticker`
is intended to act as a timeout so that block publishing is only
attempted after a given duration. However, the timeout checking (used to
trigger view changes) was also occuring in this `Ticker`; this meant
that these timeouts were only ever being checked when the block
duration had passed. These should not be tied together, so this commit
separates them and gives the ticker a more appropriate name.

Signed-off-by: Logan Seeley <seeley@bitwise.io>