Rename message_timeout and make local config
Rename the `message_timeout` config value to `update_recv_timeout` to be
more self-explanatory.

Also removes this value as an on-chain setting and instead makes it a
locally configured value (via CLI argument); this timeout is specific to
the local consensus engine <-> validator connection, so it should not be
configured with a network-wide setting.

Signed-off-by: Logan Seeley <seeley@bitwise.io>