Rename storage and make local config
Rename the `storage` config value to `storage_location` to be more
self-explanatory.

Also removes this value as an on-chain setting and instead makes it a
locally configured value (via CLI argument); this setting is
deployment-specific and it is possible for nodes with different storage
types/locations to exist on the same network, so it should not be
configured with a network-wide setting.

Signed-off-by: Logan Seeley <seeley@bitwise.io>