On view change, old primary should cancel block
When a view change is performed, the old primary should cancel any block
it may have initialized; if/when it becomes primary again in the future,
it will need initialize a new block that is built on top of the new
chain head.

This fixes a bug where nodes would get stuck because an old intialized
block was never cancelled.

Signed-off-by: Logan Seeley <seeley@bitwise.io>