Fail old blocks
Updates the `on_block_new` method to fail any blocks that are older than
the current sequence number. This was not done previously because the
Sawtooth validator would crash if consensus issued a {commit, ignore, or
fail} command for the same block more than once. This validator bug has
since been fixed, so PBFT can fail any old blocks without worrying about
the validator crashing. This is beneficial because it allows PBFT to
uphold the guarantee that it decides on all blocks it receives from the
validator.

Signed-off-by: Logan Seeley <seeley@bitwise.io>