Check blocks and handle BlockValid/BlockInvalid
Updates PBFT to check blocks that it receives the validator and handle
the resulting BlockValid and BlockInvalid messages. This change is
necessary due to the Sawtooth validator being updated to validate blocks
only when specifically requested by consensus.

Overall changes made to PBFT:
- When PBFT receives a block and does its verification, it calls
`service.check_blocks` to validate the block and adds the block to the
log as an unvalidated block.
- BlockInvalid messages are handled by a new `on_block_invalid` method.
This method drops the block from the log and fails it.
- BlockValid messages are handled by a new `on_block_valid` method. This
method marks the block as valid in the log then processes the block as
usual (sends a PrePrepare, catches up, etc.).

Changes made to the log:
- Log has a new HashMap that stores unvalidated blocks along with their
corresponding seals.
- The `add_block` method is split into two methods:
`add_validated_block` which is used to add blocks that have already been
validated (such as chain head on startup) and `add_unvalidated_block`
which is used to add blocks that have not yet been validated (when
received as a BlockNew).
- `block_validated` and `block_invalidated` methods are added to move a
block from `unvalidated_blocks` to `blocks` or drop a block,
respectively.

Signed-off-by: Logan Seeley <seeley@bitwise.io>