Move seal validation after block validation
Updates the block verification procedure to have the node check the
validity of the block itself (from the validator's perspective) before
checking the validity of the block's seal. This will simplify the
process of waiting for the appropriate blocks to be committed before
verifying a seal, and also removes the need to store an unvalidated
block's seal in the log.

Signed-off-by: Logan Seeley <seeley@bitwise.io>