Only handle block when grandparent is committed
Updates the consensus seal verification and block handling procedure to
only verify the seal and handle the block when its grandparent has been
committed. A block's grandparent must be committed before verifying the
seal of a block, since the settings in state as of the grandparent are
required for validating the block's seal.

Signed-off-by: Logan Seeley <seeley@bitwise.io>