Prevent bootstrap commits from being ignored
If on a complete restart f nodes are out of synch, and
if the node start is staggered so that a view change is
initiated before all the bootstrap commits are received
then the network can get into a situation where it never
starts.  This change allows the commits into the message
log during ViewChange if they are for the already committed
chain head. This is safe, since obviously the block is
already committed.

Signed-off-by: Kevin O'Donnell <kevin@blockchaintp.com>