Pin PBFT liveness tests to 1-3 nightly validator
sawtooth-core is currently working toward 2.0 and may not
be compatibly with this current version of PBFT. For now
the liveness test will continue to be run against 1-3,
the "main" branch for further 1.x develpment.

Also 1-3 does not include the python3-sawtooth-settings, so it
has been replaced with the rust version, sawtooth-settings-tp.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>